cleanup maps and layers
